+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ### 2016-01-20 Version 1.0.3
   * Bug fix: hand compute interactions in `sem.lavaan` using argument `compute.int = TRUE`
-  * Feature addition: function `scale.data` to handle transformed variables in `sem.coefs` when standardize = "scale"
+  * Feature addition: function `get.scaled.data` to handle transformed variables in `sem.coefs` when standardize = "scale"
 
 ### 2016-01-15 Version 1.0.2
   * Feature addition: added additional plotting arguments for partial.residuals